Changes
Today I deleted 3 images, added 1 images, changes a name of a file and
changes 3 different files. Inside the generic I was starting to add the
essay to the paragraphs. Inside the  index I was also adding
unnecessary code and added to the paragraph. Finally  I added the
caption to the style sheet.